+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this library are documented in this file.
 
 - Add `nwsice` color ramp per suggestion.
 - Retain reference to `CGIModel` schema as `environ["_cgimodel_schema"]`.
+- Parse flight level within PIREPs (#1003).
 - Use `akrherz/iem_database:test_data` container for integration tests.
 
 ### Bug Fixes